requirements.txt
accelerate==1.0.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
airportsdata==20241001
annotated-types==0.7.0
anyio==4.7.0
astor==0.8.1
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.3.0
auto_gptq==0.7.1
autoawq==0.2.7.post3
blake3==1.0.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
coloredlogs==15.0.1
comm==0.2.2
compressed-tensors==0.8.1
contourpy==1.3.0
cycler==0.12.1
datasets==3.2.0
debugpy==1.8.6
decorator==5.1.1
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
gekko==1.2.1
gguf==0.10.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
inquirerpy==0.3.4
interegular==0.3.3
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lark==1.2.2
lm-format-enforcer==0.10.9
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistral_common==1.5.1
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.18.6
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.58.1
opencv-python-headless==4.10.0.84
optimum==1.23.3
outlines==0.1.11
outlines_core==0.1.26
packaging==24.1
pandas==2.2.3
parso==0.8.4
partial-json-parser==0.2.1.1.post4
peft==0.14.0
pexpect==4.9.0
pfzy==0.3.4
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==18.1.0
pybind11==2.13.6
pycountry==24.6.1
pydantic==2.10.3
pydantic_core==2.27.1
Pygments==2.18.0
pyparsing==3.1.4
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.40.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rouge==1.0.1
rpds-py==0.22.3
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.13.1
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
wcwidth==0.2.13
websockets==14.1
xformers==0.0.28.post3
xgrammar==0.1.7
xxhash==3.5.0
yarl==1.18.3
zipp==3.20.2
zstandard==0.23.0
